move wireguard logic to api call
